\name{getDefaultStartCodon}
\alias{getDefaultStartCodon}
\title{
Get the Default Start Codon List
}
\description{
The default start codon list will be "AUG", "AUC", "AUU", "AUA", "AGG", "ACG", 
        "AAG", "CUG", "UUG", and "GUG".
}
\usage{
getDefaultStartCodon()
}
\value{
Character vector, the default set of start codons.
}
\keyword{ methods }
