\name{getReversedSequence}
\alias{getReversedSequence}
\title{
Reverse a Fragment of DNA/RNA Sequence
}
\description{
Simply reverse a fragment of sequence (character vector).
}
\usage{
getReversedSequence(seq_fragment)
}
\arguments{
  \item{seq_fragment}{
  Character vector, a fragment of DNA/RNA sequence.
}
}
\value{
Character vector, same bases as input sequences but in reversed order.
}
\keyword{ methods }
