\name{plotMultiMetageneCounts}
\alias{plotMultiMetageneCounts}
\title{
Plot Reads Counts on Metagene Positions
}
\description{
Generate a line plot to show reads counts on each defined metagene position for multiple samples.
}
\usage{
plotMultiMetageneCounts(count_files, from_position=-50, to_position=50, 
	codon="Start", x_interval=10, line_colors = c("red", "blue", "green"))
}
\arguments{
  \item{count_files}{
  Character vector, names (and path) of metagene reads count files generated by FpFraming().
}
  \item{from_position}{
  Integer, start position of metagene to plot.
}
  \item{to_position}{
  Integer, stop position of metagene to plot.
}
  \item{codon}{
  Character vector, codon name, either "Start" or "Stop" for x-axis labeling.
}
  \item{x_interval}{
  Positive integer, interval for metagene position lables on x-axis.
}
  \item{line_colors}{
  Character vector of R color names for lines.
}
}
\author{
Henry Zhang
}
\keyword{ methods }

