\name{readAsiteFromFile}
\alias{readAsiteFromFile}
\title{
Read a_site from A-site file
}
\description{
Read a-site table generated by FpFraming(), covert it to a one column data frame with required column name.
}
\usage{
readAsiteFromFile(a_site_file)
}
\arguments{
  \item{a_site_file}{
Characte vector, name (and path) of a-site file.
}
}
\value{
A data frame with one column for a-site of each read length. Read length are used for row names.
}
\references{
Nicholas T. Ingolia, et al. (2014). Ribosome Profiling Reveals Pervasive Translation Outside of Annotated Protein-Coding Genes. Cell Reports 8, 1365-1379.
}
\keyword{ methods }
