\name{RiboProTools-package}
\alias{RiboProTools-package}
\alias{RiboProTools}
\docType{package}
\title{
R Tools for Ribosomal Footprints Profiling Analysis
}
\description{
A list of methods for Ribosomal footprints (fp) profiling analysis 
was provided including of: fp framing, fp counting, DESeq2 analysis, 
and data visualization.
}
\details{
Package: RiboProTools\cr
Type: Package \cr
Version: 1.0.0 \cr
Date: 2018-8-08 \cr
License: GPL (>=2) \cr
}
\author{
Hongen Zhang
Fujun Zhou
Maintainer: Hongen Zhang <hzhang@mail.nih.gov>
}
\keyword{ package }
