\name{validateParameters}
\alias{validateParameters}
\title{
Validate All Parameters of FpFraming
}
\description{
Validate all user defined parameters for FpFraming.
}
\usage{
validateParameters(parameters)
}
\arguments{
  \item{parameters}{
  parameters, list of numeric variables.
}
}
\value{
None.
}
\keyword{ methods }
