\name{yeast_gene_description}
\alias{yeast_gene_description}
\docType{data}
\title{
Yeast Gene Description
}
\description{
A data Frame with three columns for gene ID, gene name, and gene description.
}
\usage{data("yeast_gene_description")}
\format{
  A data frame with 7133 observations on the following 3 variables.
  \describe{
    \item{\code{V1}}{a character vector}
    \item{\code{V2}}{a character vector}
    \item{\code{V3}}{a character vector}
  }
}
\examples{
data(yeast_gene_description)
}
\keyword{datasets}
